Refactor : 파일 정리 및 성능 개선
